poetry-demo/pyproject.toml
[project]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "LuaYiNing",email = "luayining@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "ruff (>=0.9.6,<0.10.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

